+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/positioning/PositionApis.ts
@@ -1,8 +1,9 @@
 import { ValueSchema } from '@ephox/boulder';
-import { window } from '@ephox/dom-globals';
+import { document } from '@ephox/dom-globals';
 import { Adt, Fun, Option } from '@ephox/katamari';
 import { Css, Element, Location } from '@ephox/sugar';
 
+import * as AriaFocus from '../../alien/AriaFocus';
 import { Bounds, box } from '../../alien/Boxes';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { Stateless } from '../../behaviour/common/BehaviourState';
@@ -16,7 +17,10 @@ import { PositioningConfig } from './PositioningTypes';
 export interface OriginAdt extends Adt { }
 
 const getFixedOrigin = (): OriginAdt => {
-  return Origins.fixed(0, 0, window.innerWidth, window.innerHeight);
+  // Don't use window.innerWidth/innerHeight here, as we don't want to include scrollbars
+  // since the right/bottom position is based on the edge of the scrollbar not the window
+  const html = document.documentElement;
+  return Origins.fixed(0, 0, html.clientWidth, html.clientHeight);
 };
 
 const getRelativeOrigin = (component: AlloyComponent): OriginAdt => {
@@ -45,46 +49,48 @@ const positionWithin = (component: AlloyComponent, posConfig: PositioningConfig,
 const positionWithinBounds = (component: AlloyComponent, posConfig: PositioningConfig, posState: Stateless, anchor: AnchorSpec, placee: AlloyComponent, bounds: Option<Bounds>): void => {
   const anchorage: AnchorDetail<any> = ValueSchema.asRawOrDie('positioning anchor.info', AnchorSchema, anchor);
 
-  // We set it to be fixed, so that it doesn't interfere with the layout of anything
-  // when calculating anchors
-  Css.set(placee.element(), 'position', 'fixed');
+  // Preserve the focus as IE 11 loses it when setting visibility to hidden
+  AriaFocus.preserve(() => {
+    // We set it to be fixed, so that it doesn't interfere with the layout of anything
+    // when calculating anchors
+    Css.set(placee.element(), 'position', 'fixed');
 
-  const oldVisibility = Css.getRaw(placee.element(), 'visibility');
-  // INVESTIGATE: Will hiding the popup cause issues for focus?
-  Css.set(placee.element(), 'visibility', 'hidden');
+    const oldVisibility = Css.getRaw(placee.element(), 'visibility');
+    Css.set(placee.element(), 'visibility', 'hidden');
 
-  // We need to calculate the origin (esp. the bounding client rect) *after* we have done
-  // all the preprocessing of the component and placee. Otherwise, the relative positions
-  // (bottom and right) will be using the wrong dimensions
-  const origin = posConfig.useFixed ? getFixedOrigin() : getRelativeOrigin(component);
+    // We need to calculate the origin (esp. the bounding client rect) *after* we have done
+    // all the preprocessing of the component and placee. Otherwise, the relative positions
+    // (bottom and right) will be using the wrong dimensions
+    const origin = posConfig.useFixed() ? getFixedOrigin() : getRelativeOrigin(component);
 
-  const placer = anchorage.placement;
+    const placer = anchorage.placement;
 
-  const getBounds = bounds.map(Fun.constant).or(posConfig.getBounds);
+    const getBounds = bounds.map(Fun.constant).or(posConfig.getBounds);
 
-  placer(component, anchorage, origin).each((anchoring) => {
-    const doPlace = anchoring.placer.getOr(place);
-    doPlace(component, origin, anchoring, getBounds, placee);
-  });
+    placer(component, anchorage, origin).each((anchoring) => {
+      const doPlace = anchoring.placer.getOr(place);
+      doPlace(component, origin, anchoring, getBounds, placee);
+    });
 
-  oldVisibility.fold(() => {
-    Css.remove(placee.element(), 'visibility');
-  }, (vis) => {
-    Css.set(placee.element(), 'visibility', vis);
-  });
+    oldVisibility.fold(() => {
+      Css.remove(placee.element(), 'visibility');
+    }, (vis) => {
+      Css.set(placee.element(), 'visibility', vis);
+    });
 
-  // We need to remove position: fixed put on by above code if it is not needed.
-  if (
-    Css.getRaw(placee.element(), 'left').isNone() &&
-    Css.getRaw(placee.element(), 'top').isNone() &&
-    Css.getRaw(placee.element(), 'right').isNone() &&
-    Css.getRaw(placee.element(), 'bottom').isNone() &&
-    Css.getRaw(placee.element(), 'position').is('fixed')
-  ) { Css.remove(placee.element(), 'position'); }
+    // We need to remove position: fixed put on by above code if it is not needed.
+    if (
+      Css.getRaw(placee.element(), 'left').isNone() &&
+      Css.getRaw(placee.element(), 'top').isNone() &&
+      Css.getRaw(placee.element(), 'right').isNone() &&
+      Css.getRaw(placee.element(), 'bottom').isNone() &&
+      Css.getRaw(placee.element(), 'position').is('fixed')
+    ) { Css.remove(placee.element(), 'position'); }
+  }, placee.element());
 };
 
 const getMode = (component: AlloyComponent, pConfig: PositioningConfig, pState: Stateless): string => {
-  return pConfig.useFixed ? 'fixed' : 'absolute';
+  return pConfig.useFixed() ? 'fixed' : 'absolute';
 };
 
 export {