+++ modules/alloy/src/main/ts/ephox/alloy/positioning/mode/MakeshiftAnchor.ts
@@ -6,12 +6,13 @@ import { AlloyComponent } from '../../api/component/ComponentApi';
 import * as Fields from '../../data/Fields';
 import * as Bubble from '../layout/Bubble';
 import * as Layout from '../layout/Layout';
-import { OriginAdt } from '../layout/Origins';
+import * as Origins from '../layout/Origins';
 import { MakeshiftAnchor, nu as NuAnchoring } from './Anchoring';
 import * as AnchorLayouts from './AnchorLayouts';
 
-const placement = (component: AlloyComponent, anchorInfo: MakeshiftAnchor, origin: OriginAdt) => {
-  const anchorBox = bounds(anchorInfo.x, anchorInfo.y, anchorInfo.width, anchorInfo.height);
+const placement = (component: AlloyComponent, anchorInfo: MakeshiftAnchor, origin: Origins.OriginAdt) => {
+  const pos = Origins.translate(origin, anchorInfo.x, anchorInfo.y);
+  const anchorBox = bounds(pos.left(), pos.top(), anchorInfo.width, anchorInfo.height);
 
   const layouts = AnchorLayouts.get(component.element(), anchorInfo, Layout.all(), Layout.allRtl());