+++ modules/alloy/src/main/ts/ephox/alloy/ui/types/SplitFloatingToolbarTypes.ts
@@ -2,7 +2,7 @@ import { LazySink } from '../../api/component/CommonTypes';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { SimpleOrSketchSpec } from '../../api/component/SpecTypes';
 import { AnchorSpec } from '../../positioning/mode/Anchoring';
-import { SplitToolbarBaseDetail, SplitToolbarBaseSpec, SplitToolbarBaseSketcher } from './SplitToolbarBaseTypes';
+import { SplitToolbarBaseDetail, SplitToolbarBaseSpec, SplitToolbarBaseSketcher, SplitToolbarBaseApis } from './SplitToolbarBaseTypes';
 import { ToolbarGroupSpec } from './ToolbarGroupTypes';
 import { ToolbarSpec } from './ToolbarTypes';
 
@@ -13,6 +13,10 @@ export interface SplitFloatingToolbarDetail extends SplitToolbarBaseDetail {
   markers: {
     overflowToggledClass: string;
   };
+}
+
+export interface SplitFloatingToolbarApis extends SplitToolbarBaseApis {
+  reposition: (toolbar: AlloyComponent) => void;
 }
 
 export interface SplitFloatingToolbarSpec extends SplitToolbarBaseSpec {
@@ -30,4 +34,4 @@ export interface SplitFloatingToolbarSpec extends SplitToolbarBaseSpec {
   };
 }
 
-export interface SplitFloatingToolbarSketcher extends SplitToolbarBaseSketcher<SplitFloatingToolbarSpec, SplitFloatingToolbarDetail> { }
+export interface SplitFloatingToolbarSketcher extends SplitToolbarBaseSketcher<SplitFloatingToolbarSpec, SplitFloatingToolbarDetail>, SplitFloatingToolbarApis { }